# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mezquita Al-Quds API",
        "description": "API pública y administrativa de la Mezquita Al-Quds. Horarios de oración, noticias, eventos y donativos.",
        "contact": {
            "email": "info@mezquita-alquds.cat"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://mezquita-alquds.cat/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/noticias": {
            "get": {
                "tags": [
                    "Público"
                ],
                "summary": "Listar noticias publicadas",
                "operationId": "ec806f2451daeafe13d20ffcee0702d4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de noticias"
                    }
                }
            }
        },
        "/noticias/{noticia}": {
            "get": {
                "tags": [
                    "Público"
                ],
                "summary": "Ver detalle de noticia",
                "operationId": "3348019c010ec99cc0c621cf4a1b9e17",
                "parameters": [
                    {
                        "name": "noticia",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Noticia con comentarios"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}